🏗️ Move `UserCard` component to `src/components/Nav/` folder
This component is used only in the main menus of the app, so it's clearer to find it with the other navigation components.
